Update everything except icons for latest beta
Bumps analyzer version to fix a runtime issue when running. I added the latest versions of everything else while I was in there for consistency.

This change does not include icon files because there are a lot of files modified that don't appear to be changes that would make this change noisy, so I'll put them in their own change.