Update icons for latest beta (on macOS) (#46)
This updates the icons for the latest Flutter beta. The non-icon changes
are in https://github.com/flutter/tools_metadata/pull/45

As far as I can tell, there are no visual differences here but the files
did change in size slight.y I don't know if this is due to subtle
differences in Flutter or its dependencies or differences between
machines. On macOS there were around 1,800 changed files but if I ran
this on Windows there were 20,000 changed files.